TA05 Correccion double en /

+++ TA05/src/EX13/Ex13App.java
@@ -35,7 +35,7 @@ public class Ex13App {
 			System.out.println(num1 + " * " + num2 + " = " + (num1*num2));
 			break;
 		case '/':
-			System.out.println(num1 + " / " + num2 + " = " + (num1/num2));
+			System.out.println(num1 + " / " + num2 + " = " + ((double)num1/num2));
 			break;
 		case '^':
 			System.out.println(num1 + "^" + num2 + " = " + (Math.pow(num1, num2)));
